Write IsolatedSourceData peaks as int, matching the read path
`IsolatedSourceData.as_dict` previously serialized the peak tuple via
`float(p)` while `from_dict` read it back via `int(p)`. The dataclass
annotates `peak: tuple[int, int]`, and `IsolatedSource.from_footprint`
sources peaks from `getIy/getIx` (ints), so the on-disk float was an
asymmetric truncation channel for any sub-pixel peak that ever reached
this path. Switched the write to `int(p)` so the type matches the
contract and the read side.

Adds `test_as_dict_peak_is_int` asserting the encoded `peak` values
are `int`.

Audit finding IO-1 of `audits/audit-2026-05-05.md`. Pair coded with
@fred3m.

+++ python/lsst/meas/extensions/scarlet/io/source_data.py
@@ -77,7 +77,7 @@ class IsolatedSourceData(scl.io.blend.ScarletSourceBaseData):
         result: dict[str, Any] = {
             "origin": tuple(int(o) for o in self.origin),
             "shape": tuple(int(s) for s in self.span_array.shape),
-            "peak": tuple(float(p) for p in self.peak),
+            "peak": tuple(int(p) for p in self.peak),
             "span_array": tuple(self.span_array.flatten().astype(float)),
             "source_type": self.source_type,
             "version": self.version,
